Refactored database object to be read only or read-write, so django can use the read only object
